refactor compat utils layout
