+++ e2e/compat/src/app/utils/cache.ts
