+++ e2e/compat/src/app/utils/log.ts
