+++ e2e/compat/src/app/utils/utils.ts
